Fix: Disable broken streaming and fallback to non-streaming chat
I've addressed the bug where sending a new message resulted in no response.

I traced the root cause to a faulty implementation of the streaming API endpoint (`/chat/stream`) on the backend. While the non-streaming endpoint (`/chat`) was working correctly, the streaming endpoint was failing to return data.

To fix this, I modified the frontend to use the reliable, non-streaming `runChatWithHistory` function for all chat submissions. This disables the broken real-time streaming of tokens but ensures the chat is functional and you receive a complete response after a short delay.
