+++ frontend.log
@@ -3,7 +3,7 @@
 > vite
 
 
-  VITE v7.1.1  ready in 418 ms
+  VITE v7.1.2  ready in 248 ms
 
   ➜  Local:   http://localhost:5173/
   ➜  Network: use --host to expose
