+++ server/config.py
@@ -1,0 +1,11 @@
+from pydantic_settings import BaseSettings
+from dotenv import load_dotenv
+
+load_dotenv()
+
+class Settings(BaseSettings):
+    GOOGLE_API_KEY: str | None = None
+    TAVILY_API_KEY: str | None = None
+    OPENROUTER_API_KEY: str | None = None
+
+settings = Settings()
